refactor: move keybindings to it's route
