+++ apps/core/src/features/keybindings/action-utils.tsx
@@ -1,6 +1,13 @@
 import type { Row } from "@tanstack/react-table"
 import { writeText } from "@tauri-apps/plugin-clipboard-manager"
 
+export const focusSearch = () => {
+  const inputElem = document.getElementById("search_input")
+  if (inputElem) {
+    inputElem.focus()
+  }
+}
+
 export const copyRows = async (rows: Row<any>[]) => {
   if (rows.length === 0) return
   return await writeText(
