+++ apps/core/src/features/keybindings/index.ts
@@ -1,0 +1,2 @@
+export * from "./keybindings"
+export * from "./manager"
