--- apps/core/src/keybindings/index.ts
@@ -1,6 +1,0 @@
-export const focusSearch = () => {
-  const inputElem = document.getElementById("search_input")
-  if (inputElem) {
-    inputElem.focus()
-  }
-}
