+++ apps/core/src/main.tsx
@@ -1,9 +1,9 @@
+import { KeybindingsContext, KeybindingsManager } from "@/features/keybindings"
 import { SettingsContext, SettingsManager } from "@/features/settings/manager"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 import { RouterProvider, createRouter } from "@tanstack/react-router"
 import ReactDOM from "react-dom/client"
 import "./index.css"
-import { KeybindingsContext, KeybindingsManager } from "./keybindings/manager"
 import { routeTree } from "./routeTree.gen"
 
 const client = new QueryClient()
