+++ apps/core/src/routes/dashboard/_layout/_table-view-layout/table-view/$tableName/-components/data-table.tsx
@@ -30,9 +30,9 @@ import {
 } from "@/components/ui/context-menu"
 import { ScrollArea, ScrollBar } from "@/components/ui/scroll-area"
 import { Sheet } from "@/components/ui/sheet"
+import { useKeybindings } from "@/features/keybindings"
+import { copyRows } from "@/features/keybindings/action-utils"
 import { useSetupReactTable } from "@/hooks/table"
-import { useKeybindings } from "@/keybindings/manager"
-import { copyRows } from "@/keybindings/row"
 import { useEditRowSheetState } from "@/state/sheetState"
 import { useTableState } from "@/state/tableState"
 import { useQueryClient } from "@tanstack/react-query"
