+++ apps/core/src/routes/dashboard/_layout/keybindings/route.tsx
@@ -1,5 +1,10 @@
-import { createFileRoute } from '@tanstack/react-router'
+import { Keybindings } from "@/features/keybindings"
+import { createFileRoute } from "@tanstack/react-router"
 
-export const Route = createFileRoute('/dashboard/_layout/keybindings')({
-  component: () => <div>Hello /dashboard/_layout/keybindings!</div>
+export const Route = createFileRoute("/dashboard/_layout/keybindings")({
+  component: KeybindingsRoute
 })
+
+function KeybindingsRoute() {
+  return <Keybindings />
+}
